+++ Dockerfile
@@ -32,4 +32,4 @@ RUN echo "==> Clean up..."  && \
 
 EXPOSE 7860
 
-CMD [ "python3", "app/web.py", "--host", "0.0.0.0", "--port", "7860"]
+CMD [ "python3", "app.py", "--host", "0.0.0.0", "--port", "7860"]
