+++ docker-compose.yml
@@ -5,7 +5,7 @@ services:
     build:
       context: .
       dockerfile: Dockerfile
-    image: hivision_idphotos
-    command: python3 web.py --host 0.0.0.0 --port 7860
+    image: hivision_idphotos2
+    command: python3 app.py --host 0.0.0.0 --port 7860
     ports:
       - "7860:7860"
